\name{pSI-package}
\alias{pSI-package}
\alias{pSI}
\docType{package}
\title{
Specific Index Statistic
}
\description{
This package contains functions to calculate the Specificity Index statistic, which can be used for comparative quantitative analysis to identify genes enriched in specific cell populations across a large number of profiles, as well as perform numerous post-processing operations

NOTE:Supplementary data (supplementary tables,human & mouse expression sets, calculated pSI datasets, candidate gene lists, etc.) can be found in \code{pSI.data} package.
}
\details{
\tabular{ll}{
Package: \tab pSI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-03\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Xiaoxiao Xu, Alan B. Wells, David OBrien, Arye Nehorai, Joseph D. Dougherty

Maintainer:Alan B. Wells <awells@dsgmail.wustl.edu>
}
\references{
Joseph D. Dougherty, Eric F. Schmidt, Miho Nakajima, and Nathaniel Heintz
Analytical approaches to RNA profiling data for the identification of genes enriched in specific cells
\url{http://nar.oxfordjournals.org/content/38/13/4218}
}