\docType{data}
\name{sample.data}
\alias{sample.data}
\title{Sample Input & Output pSI Data Sets}
\format{List comprised of two data frames.}
\source{
\url{http://www.brainspan.org}
}
\description{
Sample expression matrix and sample specificity index
statistic (pSI) output data sets used to illustrate the
\code{pSI} package functions.
}
\details{
Sample expression matrix is the Brainspan RNAseq data
condensed into 6 major regional divisions for just the
young adult regions. Sample pSI output data set is
calculated from the young adulthood Brainspand RNAseq data
just previously described.
}
\examples{
data(sample.data)
}
\keyword{datasets}

